backend/requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.41.0
python-dotenv==1.0.1
pydantic>=2.12.0
supabase>=2.28.0
websockets>=11,<16
httpx>=0.27
